# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Remake of my previous attampted on creating API with Node.js",
        "title": "Gin API",
        "contact": {
            "name": "API Support",
            "url": "https://mortis-grimreaper.ddns.net/projects",
            "email": "AndrewYushchenko@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/YushchenkoAndrew/API_Server/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:31337",
    "basePath": "/api",
    "paths": {
        "/info": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Read All Info",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Instance :id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "CreatedAt date",
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Countries: 'UK,US'",
                        "name": "countries",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Info"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Update Info by Query",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Instance :id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "CreatedAt date",
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Countries: 'UK,US'",
                        "name": "countries",
                        "in": "query"
                    },
                    {
                        "description": "Info Data",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReqInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Info"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "'CreatedAt' setted automatically",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Create one instace of Info",
                "parameters": [
                    {
                        "description": "Info Data",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReqInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Info"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Delete Info by Query",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Instance :id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "CreatedAt date",
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Countries: 'UK,US'",
                        "name": "countries",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/info/list": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Create Info from list of objects",
                "parameters": [
                    {
                        "description": "List of Info Data",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ReqInfo"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Info"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
        "/info/{date}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Create/Update Info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Created at instance",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Info Data",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReqInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Info"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Server side error: Something went wrong"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "ERR"
                }
            }
        },
        "models.Info": {
            "type": "object",
            "properties": {
                "Clicks": {
                    "type": "integer",
                    "example": 2
                },
                "Countries": {
                    "type": "string",
                    "example": "UK,US"
                },
                "CreatedAt": {
                    "type": "string",
                    "example": "2021-08-06"
                },
                "ID": {
                    "type": "integer"
                },
                "Media": {
                    "type": "integer",
                    "example": 3
                },
                "Views": {
                    "type": "integer",
                    "example": 1
                },
                "Visitors": {
                    "type": "integer",
                    "example": 4
                }
            }
        },
...
        "models.ReqInfo": {
            "type": "object",
            "properties": {
                "Clicks": {
                    "type": "integer"
                },
                "Countries": {
                    "description": "ID        uint32    `json:\"id\" xml:\"id\"`\nCreatedAt *time.Time `json:\"CreatedAt\" xml:\"CreatedAt\"`",
                    "type": "string"
                },
                "Media": {
                    "type": "integer"
                },
                "Views": {
                    "type": "integer"
                },
                "Visitors": {
                    "type": "integer"
                }
            }
        },
...
        "models.Success": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "integer",
                    "example": 1
                },
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "page": {
                    "type": "integer",
                    "example": 0
                },
                "result": {
                    "type": "object"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "totalItems": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
...
    }
}